Update makefile.
Re-run cbp2make including the options to tell cbp2make to not remove
the bin and xLights/.objs subdirectories when running the clean
operation.  --with-deps was also used to include dependency information
in the makefile.

The command used to create xLights.cbp.mak using cbp2make is now documented
at the bottom of README.linux as well.