Commit for filesave defect and Enhanced grid visibility
Fixed code to properly clear the xLights filename when new sequences are
created.  Restructured code to ensure that the updates to the displayed file
name happen only for the right events and that the xlights file name is only
updated when the displayed name is updated.

Enhanced effects grid so that it highlights the current cell.