pseudocode to change the 1 is unmask layerinng methid.
old method says if one is non black then allow effect 2 to shine through
new method is the same if statement but adjust the brightness of effect 2 to have the brightness value
from effect1. In other words effct2.hsv we leave the h and s alone, we change the v.