Fix Linux compile by updating clipboard code to use wxClipboard
The wxWidgets wxOpenClipboard(), wxEmptyClipboard(),
wxSetClipboardData(), wxGetClipboardData(), and wxCloseClipboard()
calls are deprecated and were only implemented on Windows according
to wxWidgets docs.  This patch converts the code to use the newer
wxClipboard class and global wxTheClipboard object to get rid of
the deprecated calls and re-enable compilation on Linux.

Some warning dialogs were also added in the process to inform users
in case their data couldn't be copied to or from the clipboard for
one of several reasons.