Fix preview window refreshing on Linux
wxWindow::ClearBackground() does not work properly on Linux so you are
left with artifacts all over the preview window when moving a model
around the screen.  This patch goes a level deeper and uses
wxClientDC::Clear() to clear the preview window instead.  Prior to
changing to wxClientDC, I also tried ::SetBackgroundColour() and that
would not clear the window either.  This appears to possibly be a
wxGTK related issue possibly based on an internet search.