Fix compiling on Linux, with a limitation.
'Button1' variable name conflicts with X #define so rename it.
'None' variable name conflicts with something else so rename it.

wxGridCellEditorEvtHandler is undefined on Linux, so ModelFaceDialog
is partially broken.  This will need a workaround.  The same still
holds true for usage of the same wx class in the Papagayo code.  I looked
into this a little when the Papagayo workaround was added, but I don't
know enough about wxWidgets to suggest a true fix for Linux users.