Remove OpenGL stuff from warp effect (#1151)
The intention here was that since this may be an expensive effect to render, a shader implementation might be faster even though that would require that the rendering be done on the main thread. I'm not sure that it was actually faster though, and it was only building successfully on Windows, so it seems like it's more trouble than it's worth at this point.