Add ability to pan in the X/Y plane using CTRL modifier.
When holding down the CTRL modifier, add ability to pan in the X/Y
plane.

Existing behavior of panning in the X/Z plane is retained when no
modifier key is down.

Adding X/Y plane panning handles the use case where the user uses X/Z
panning to move close to their house model, and then wishes to layout
the lights along a face of the house.  X/Y panning provides a means of
moving across the face of the house without having to rotate the camera
view and re-adjust panning in the X/Z plane.