Add configurable starRatio parameter for StarModel.
Stars in xLights currently use a ratio of 2.618034 between the inner and
outer radius.  Some stars commercially available from common supplies do
not have the same ratio.  Being able to change the ratio of the star
provides the user with a means to more accurately model these stars
without relying on a custom model definition.