~Added experimental Hinkspix Pro/Easylights 16 Controller Upload
I hope to clear up some misinformation.

The EasyLights code, Joe wrote in Jan 2019, used UPD sockets to setup the Easylights controller. Around August 2019, Joe rewrote the controller firmware removing the UPD support and breaking the xLights code. After trying and failing to use the easylights upload with a couple users in zoom I emailed/called Joe asking him if it was still functional. He told me he removed the UPD support from the controller and as he is not an active xLights developer, he had no plans to support that code anymore, and asked me to remove it. In an attempt to support our users, I bought a EasyLights 16 controller, and reverse engineered the controllers HTTP web interface to upload the data in xLights with HTTP GET calls. Based on the javascript code, I wrote the xLights code to work on the HinksPix Pro and EasyLights 16 controller. I have only tested it with the easylights controller I have. I may buy a HinkPix Pro in the future and test the code with it but it is hard to waste $300 on controllers I do not plan on using. If anyone runs into issues please send me the logs and I will try to work through them. I did this all on my own accord/stupidity, no one paid me, I just wanted to help our users. If anyone has any questions please let me know.

Thanks,
Scott