Fix further memory leaks (#1915)
* Fix further memory leaks

Seeing as the last one caused a few problems, I'll try
to be more explicit about the changes here.

Tracelog - The per-thread trace buffers aren't tidied up,
and some still have log messages present. These were leaking, so
they're now freed up in the destructor.

ViewObjectManager - The view objects were never removed. Invoke the
existing "clear" function from the destructor.

XLightsMain UI Elements - We're leaking button elements for some
reason - Added delete calls, matching the pattern already present for
CheckBoxLightOutput.

XLightsMain XmlSetting - Need to delete elements that we've obtained
through "RemoveChild", as per the wxWidgets docs.

Testing performed -
 - Opened layout
 - checked layout in 2d/3d
 - opened sequence
   - rendered
   - played
   - changed
   - played
   - saved
- closed xLights
- Repeated the above to verify changes saved

When I have saved a sequence, I do see an access violation on
shutdown, but that doesn't seem to be new. I see it on the current
master branch, and in local builds of released versions back to 2020.1.
Not sure if this occurs in the MinGW builds too, or if it's a VS thing,
but I'll have a poke around. Entirely understand if you don't want to
take this PR without that resolved, although I don't think it's
related to my changes.

* Fix log4cpp memory leaks, model leak

As per log4cpp docs, the returned vector is the callers responsibility.
Make sure we free it when we're done.

Further, when clearing up model data, ensure we remove the objects
before we remove all references from the lists.