cleanup: use intensityRangeStep value to reduce inline calculations
This allows each for loop iteration to simplfy offset a base value by the step value * i. For effects without an intensity range value (static intensity), the base value is instead the static intensity value and the range step is 0. This ensures regardless of steps, the product is always the static intensity value as desired.