fix: prevent zero value dataLength, fix off by one in #parseVariableHeaders while loop
This prevents dataLength from == 0. Previously the read len was checked, which by default should be at least 4. While checking 0 here is good, it should also check if the length - the normal variable header length (4) is >0. Otherwise a zero length data section would be encoded with a length of 4, and xLights would still attempt to read.

This additionally fixes an off by one error in the while loop. It previously checked that at least 5 bytes remained, but because the header size is rounded to the nearest product of 4 (and each variable header is 4 bytes), the code only needs to check for 4 remaining bytes to ensure there's at least room for another (even invalid) variable header.