fix: prevent maliciously, or corrupted, dataLength values from reading out of bounds
dataLength may contain a value longer than the actual data section. This results in it possibly reading out of bounds and into other VariableHeaders.

While there is no way for the program to recover from this state (besides skipping variable parsing), it will now at least detect and log an error to help determine when a sequence is corrupted.