lor-convert-fixes (#2140)
* fix: precalculate rampdiff and safely handle a zero value

This prevents duplicate calculations in perdiff iterations since startIntensity & endIntensity are not changed within the loop. This furthermore safely handles rampdiff == 0 values, by simplify ignoring the startIntensity & endIntensity code paths, as well as the intensity > 0 checks.

* fix: calculate rampdiff after intensity value rescale mutations

* fix: do not reset 0 value intensity/startIntensity/endIntensity to MaxIntensity

If the values read from the file, missing or exactly, equal 0 then those values should be truthfully copied. Generally speaking, 0% is a "safer" default failure mode than 100%.

* cleanup: condense duplicate LOR effect conversion iteration loops

This reduces code duplication and helps introduce a single pattern consistently used between each effect implementation.

* cleanup: inline twinklestate variable to reduce weird scoping access

* cleanup: inline twinkleperiod (400) and magic number (100) to named consts

* cleanup: use intensityRangeStep value to reduce inline calculations

This allows each for loop iteration to simplfy offset a base value by the step value * i. For effects without an intensity range value (static intensity), the base value is instead the static intensity value and the range step is 0. This ensures regardless of steps, the product is always the static intensity value as desired.

* fix: seed twinkleState with an initial random value

* fix: print warning if the EffectType cannot be converted

* cleanup: only read intensity values if frameCount > 0

* fix: ensure the last frame of twinkle effect ends with the correct intensity

* cleanup: localize MaxIntensity (now MaxLORIntensity) to its immediate usage, make const to hint at potential compiler optimizations

* cleanup: remove unused logger_base variable

* fix: make invalid LOR unit id wrapping consistent and handle upper bounds

This replaces a wrapping mechanism with clamping behavior. This ensures that when converting a sequence with an out of bounds unit ID, the unit ID will become either the lowest or highest unit ID (clamped). The result of this is that all invalid effects will be pushed into one of the two units, instead of being wrapped with +256 and ending up in effectively random unit IDs.

This also introduces a warning output with additional information to help identify when this issue occurs.

* fix: more randomly seed initial twinkleState using curchannel