FSEQ VariableHeader parsing improvements (#2143)
* cleanup: move magic number (4) to named const, VariableHeaderLength

* fix: prevent reading unrecongized VariableHeaders from fseq files

Each VariableHeader uses a uint16 for its length. This allows each VariableHeader's data allocation to be up to 65k. If a VariableHeader code is not recongized, this means xLights will load and maintain a useless (and fairly large) allocation.

* cleanup: merge V2FSEQ_VARIABLE_HEADER_SIZE & VariableHeaderLength to V2FSEQ_VARIABLE_HEADER_SIZE

* fix: prevent zero value dataLength, fix off by one in #parseVariableHeaders while loop

This prevents dataLength from == 0. Previously the read len was checked, which by default should be at least 4. While checking 0 here is good, it should also check if the length - the normal variable header length (4) is >0. Otherwise a zero length data section would be encoded with a length of 4, and xLights would still attempt to read.

This additionally fixes an off by one error in the while loop. It previously checked that at least 5 bytes remained, but because the header size is rounded to the nearest product of 4 (and each variable header is 4 bytes), the code only needs to check for 4 remaining bytes to ensure there's at least room for another (even invalid) variable header.

* fix: prevent maliciously, or corrupted, dataLength values from reading out of bounds

dataLength may contain a value longer than the actual data section. This results in it possibly reading out of bounds and into other VariableHeaders.

While there is no way for the program to recover from this state (besides skipping variable parsing), it will now at least detect and log an error to help determine when a sequence is corrupted.

* fix: log an error if VariableHeader's data is not null terminated

This is to assist in debugging string related issues. Non-null terminated data is strange (for the current mf & sp usages), but not really a problem.

* fix: fail if dataLength <= 0, not only == 0

If the 2 byte length value read for the VariableHeader has a value of 0 (seen in legacy code implementation), then it will equal -4 when offset by FSEQ_VARIABLE_HEADER_SIZE. This would prevent the !dataLength check from detecting the error.

* fix: always read unrecognized VariableHeaders, avoid data loss

NULL termination warnings for VariableHeader data payloads is now only printed for recognized VariableHeaders, since otherwise the data format may be entirely arbitrary.