Update ShaderEffect.cpp
A better solution, no duplication of offset control functionality, and zoom is accounted for separately.  Shaders implementing offset, but not zoom will have this added.  Caveat: the detection looks for ‘zoom ‘ in the name of the input.  If it’s called something else, the control functionality will be duplicated.  We may wish to add ‘scale,’ to the detector.