controllers:  Add support for BBB16v2
Scott Hanson's BBB16v2 has an extra 4 ports available using the diff port.
Ports 17 to 20 do not appear in the xLights visualizer so it is
not possible to assign a model to those ports.

This commit adds a controller definition
that includes the additional ports available
using the diff port.