wxSpinCtrl:  work around macOS issue
The wxWidget wxSpinCtrl() checks value in the spinner
against the min and max continuously.  If you attempt
to key a value directly and the min value is greater
than 1, it prevents the user from entering a value.

This commit fixes four instances of this issue in xLights by
changing the min value to 1 and performing min value
checking when wxSpinCtrl returns a value.