Enhancement: When a ruler is defined, show the X/Y/Z coordinates of models in the models export.
Purpose: Many users who take advantage of the 3d view in xLights use CAD or photogrammetry models of their house to layout their displays.  A user may wish to lay out their display entirely in xLights and then translate the position of models to the real world for setup.  If the user adjusts their display so that the world origin in xLights corresponds to a known reference point in the real world (e.g. the corner of a lot) the real world coordinates of the props in xLights can be used to lay out their display.  Exporting the X/Y/Z coordinates of models helps when laying out the props in the real world.

Change: When the user has defined a ruler to map between the xLights coordinate system and a known distance in the real world, add 3 new columns to the model export spreadsheet containing the "real world" X/Y/Z coorindates of each model.